requirements-dev.txt
# Development dependencies for 254Carbon Access Layer

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Database and caching
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication and security
python-jose[cryptography]==3.3.0
cryptography>=41.0.0,<47.0.0
passlib[bcrypt]==1.7.4

# WebSocket and streaming
websockets==12.0
kafka-python==2.0.2

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0

# Metrics and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
pytest-timeout==2.2.0
httpx==0.25.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Development tools
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0

# Build and deployment
docker==6.1.3
pyyaml==6.0.1
jinja2==3.1.2

# Security scanning
bandit==1.7.5
safety==2.3.5

# Performance testing
locust==2.17.0

# Development dependencies for all services in the Access Layer
# Keep versions pinned or bounded to ensure reproducible CI.
